+++ inbm/Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - ?
 
+### Added
+ - RTC 530033 - Add AOTA Applicaiton Update command to INBC
+
 ## 4.0.1 - 2023-05-26
 
 ### Security